comecei a logica do global state
