+++ src/pages/HomePage/styled.jsx
@@ -1,0 +1,18 @@
+import styled from "styled-components";
+import HomePage from "./HomePage";
+
+
+export const HomePageContainer = styled.main`
+    margin: 0 auto;
+    width: 90vw;
+    max-width: 400px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+`
+
+export const CategoryFilter = styled.div`
+    display:flex;
+
+`
